adapt to the move of infrastructure from opm-common to opm-material
