+++ opm/simulators/flow_ebos_gasoil.cpp
@@ -21,7 +21,7 @@
 
 #include <opm/simulators/flow_ebos_gasoil.hpp>
 
-#include <opm/common/utility/ResetLocale.hpp>
+#include <opm/material/common/ResetLocale.hpp>
 #include <ewoms/models/blackoil/blackoiltwophaseindices.hh>
 
 #include <dune/grid/CpGrid.hpp>
